% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cEPU_GR.R
\name{get_cEPU_GR}
\alias{get_cEPU_GR}
\title{Categorical Greek economic policy uncertainty index}
\usage{
get_cEPU_GR()
}
\value{
A data frame containing categorical Greek EPU index.
}
\description{
Extract categorical Greek economic policy uncertainty index.
}
