% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cEPU_JP.R
\name{get_cEPU_JP}
\alias{get_cEPU_JP}
\title{Categorical Japanese economic policy uncertainty index}
\usage{
get_cEPU_JP()
}
\value{
A data frame containing categorical Japanese EPU index.
}
\description{
Extract categorical Japanese economic policy uncertainty index several countries.
}
