% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cEPU_KR.R
\name{get_cEPU_KR}
\alias{get_cEPU_KR}
\title{Categorical Korean economic policy uncertainty index}
\usage{
get_cEPU_KR()
}
\value{
A data frame containing categorical Korean EPU index.
}
\description{
Extract categorical Korean economic policy uncertainty index.
}
