% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cEPU_US.R
\name{get_cEPU_US}
\alias{get_cEPU_US}
\title{Categorical American economic policy uncertainty index}
\usage{
get_cEPU_US()
}
\value{
A data frame containing categorical American EPU index.
}
\description{
Extract categorical American economic policy uncertainty index.
}
