% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CPU_US.R
\name{get_CPU_US}
\alias{get_CPU_US}
\title{American climate policy uncertainty index.}
\usage{
get_CPU_US()
}
\value{
A data frame containing American climate policy uncertainty index.
}
\description{
Extract American climate policy uncertainty index.
}
