% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EMV_US.R
\name{get_EMV_US}
\alias{get_EMV_US}
\title{American equity market volatility index}
\usage{
get_EMV_US()
}
\value{
A data frame containing American equity market volatility index.
}
\description{
Extract American equity market volatility index.
}
