% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EPU.R
\name{get_EPU}
\alias{get_EPU}
\title{Economic policy uncertainty indices}
\usage{
get_EPU()
}
\value{
A data frame containing EPU indices.
}
\description{
Extract economic policy uncertainty indices of several countries.
}
