% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EPU_JP.R
\name{get_EPU_JP}
\alias{get_EPU_JP}
\title{Japanese economic policy uncertainty index}
\usage{
get_EPU_JP()
}
\value{
A data frame containing Japanese EPU index.
}
\description{
Extract Japanese economic policy uncertainty index several countries.
}
