% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EPU_KR.R
\name{get_EPU_KR}
\alias{get_EPU_KR}
\title{Korean economic policy uncertainty index}
\usage{
get_EPU_KR()
}
\value{
A data frame containing Korean EPU index.
}
\description{
Extract Korean economic policy uncertainty index.
}
