% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EU_BR.R
\name{get_EU_BR}
\alias{get_EU_BR}
\title{Brazilian economic uncertainty index}
\usage{
get_EU_BR()
}
\value{
A data frame containing Brazilian economic uncertainty index.
}
\description{
Extract Brazilian economic uncertainty index.
}
