% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FSI_US.R
\name{get_FSI_US}
\alias{get_FSI_US}
\title{American financial stress indicator.}
\usage{
get_FSI_US()
}
\value{
A data frame containing American financial stress indicator.
}
\description{
Extract American financial stress indicator.
}
