% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GPRNK.R
\name{get_GPRNK}
\alias{get_GPRNK}
\title{Geopolitical risk from North Korea index}
\usage{
get_GPRNK()
}
\value{
A data frame containing geopolitical risk from North Korea index.
}
\description{
Extract geopolitical risk from North Korea index.
}
