% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MF_EPU.R
\name{get_MF_EPU}
\alias{get_MF_EPU}
\title{Migration fears and EPU}
\usage{
get_MF_EPU()
}
\value{
A data frame containing migration fears and EPU indices.
}
\description{
Extract migration fears and economic policy uncertainty indices of France, Germany, U.K and the U.S.
}
