% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MPU_US.R
\name{get_MPU_US}
\alias{get_MPU_US}
\title{American monetary policy uncertainty index.}
\usage{
get_MPU_US()
}
\value{
A data frame containing American monetary policy uncertainty index.
}
\description{
Extract American monetary policy uncertainty index.
}
