% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sEPU_US.R
\name{get_sEPU_US}
\alias{get_sEPU_US}
\title{State-Level American economic policy uncertainty index}
\usage{
get_sEPU_US()
}
\value{
A data frame containing State-Level American EPU index.
}
\description{
Extract State-Level American economic policy uncertainty index.
}
