% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\docType{data}
\name{references}
\alias{references}
\title{References of EPU indices}
\format{
A data frame with 38 rows and 4 columns.
}
\source{
\url{www.PolicyUncertainty.com}
}
\usage{
references
}
\description{
A dataset containing the refrences of EPU indices.
}
\keyword{datasets}
